Bugfix/expose typeID to support wrapped types (#137)
* expose typeID

* expose typeID in Kind

+++ src/Data/Morpheus/Kind/GQLType.hs
@@ -97,7 +97,7 @@ instance GQLType Int where
   typeID _ = "Int"
 
 instance GQLType Float where
-  typeID _ = "Int"
+  typeID _ = "Float"
 
 instance GQLType Text where
   typeID _ = "String"
